+++ tests/core/version_test.cpp
@@ -3,6 +3,6 @@
 #include <gtest/gtest.h>
 
 TEST(Version, IsNonEmpty) {
-    EXPECT_STREQ(torrin::kVersion, "0.3.0");
+    EXPECT_STREQ(torrin::kVersion, "1.0.0");
     EXPECT_STREQ(torrin::kName, "Torrin");
 }